+++ bindings/deno/test/mod_test.ts
@@ -13,7 +13,7 @@ import {
   MorpheusPartOfSpeech,
   MorpheusStatus,
   MorpheusTruncatedField,
-} from "./mod.ts";
+} from "../mod.ts";
 
 function assert(condition: unknown, message: string): asserts condition {
   if (!condition) throw new Error(message);